backend/requirements.txt
appdirs==1.4.4
asgiref==3.4.1
distlib==0.3.2
Django==3.2.5
filelock==3.0.12
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.4.7
djangorestframework==3.12.4
gunicorn==20.1.0
django-heroku==0.3.1
django-cors-headers==3.7.0
Pillow==8.3.1